requirements.txt
discord.py[voice]>=2.3.0
yfinance>=0.2.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
beautifulsoup4>=4.12.0
requests>=2.31.0
deep-translator>=1.11.0
scipy>=1.10.0
python-dotenv>=1.0.0
